refactor(gui): refactor all the view helper functions
